Upload rageface operator
RageFace operator""_ლಠ益ಠლ(unsigned long long n)

Usage:  This has a use?

Note that I can't build this in anything but cl.exe /utf-8 without escaping the unicode and defeating the purpose.  
I'm sure somebody will figure it out.  GCC can supposedly do it as of later version,  and I could dig around in the clang code and find a way...  but why.